+++ variables.tf
@@ -77,7 +77,7 @@ variable "console_severity" {
 }
 
 variable "destinations" {
-  description = "List of destinations. Allowed values `port`: 0-65535. Default value `port`: 514. Choices `format`: `aci`, `nxos`. Default value `format`: `aci`. Choices `facility`: `local0`, `local1` ,`local2` ,`local3` ,`local4` ,`local5`, `local6`, `local7`. Default value `facility`: `local7`. Choices `severity`: `emergencies`, `alerts`, `critical`, `errors`, `warnings`, `notifications`, `information`, `debugging`. Default value `severity`: `warnings`."
+  description = "List of destinations. Allowed values `port`: 0-65535. Default value `port`: 514. Choices `format`: `aci`, `nxos`. Default value `format`: `aci`. Choices `facility`: `local0`, `local1` ,`local2` ,`local3` ,`local4` ,`local5`, `local6`, `local7`. Default value `facility`: `local7`. Choices `severity`: `emergencies`, `alerts`, `critical`, `errors`, `warnings`, `notifications`, `information`, `debugging`. Default value `severity`: `warnings`. Choices `mgmt_epg_type`: `inb`, `oob`. Default value `mgmt_epg_type`: `inb`."
   type = list(object({
     hostname_ip   = string
     port          = optional(number)
@@ -85,7 +85,7 @@ variable "destinations" {
     format        = optional(string)
     facility      = optional(string)
     severity      = optional(string)
-    mgmt_epg      = optional(string)
+    mgmt_epg_type = optional(string)
     mgmt_epg_name = optional(string)
   }))
   default = []
@@ -127,9 +127,9 @@ variable "destinations" {
 
   validation {
     condition = alltrue([
-      for d in var.destinations : d.mgmt_epg == null || try(contains(["inb", "oob"], d.mgmt_epg), false)
+      for d in var.destinations : d.mgmt_epg_type == null || try(contains(["inb", "oob"], d.mgmt_epg_type), false)
     ])
-    error_message = "`mgmt_epg`: Allowed values are `inb` or `oob`."
+    error_message = "`mgmt_epg_type`: Allowed values are `inb` or `oob`."
   }
 
   validation {