+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- BREAKING CHANGE: rename module to syslog-policy
+- BREAKING CHANGE: rename mgmt_epg variable to mgmt_epg_type
+
 ## 0.0.1
 
 - Initial release